Move command discovery to the service manager.
